journal: continue if some files could not be opened, but warn
When trying to open journal directories:
- continue if some files could be opened, fail if no files were opened;
- emit a LOG_WARN message for all files that could not be opened for
  reason other then permission errors;
- continue to emit the hint about ACLs if they are missing

The change in behaviour is that warnings for non-EACCES errors are
more visible, but it is still possible to continue. The assumption
behind this is that we should be resilient to some files being broken,
but the user most likely wants to know about this because it indicates
some serious problems.

Only one error is printed in check_access(), because with the change
to log each error when opening files at warning level there is no need
to repeat all of them.

Fixes #1669.