boot: try to load attempts file for every entry
We always look for file named "attempts" in the same directory where loader is
stored for the given entry. If present, we load number of boot attempts
remaining.  Every time we try to boot the entry we decrement counter by one. If
an entry is selected but attempts count is zero we refuse to boot it and proceed
with next one in the list. If the entry has no attempts file associated with it,
we will consider this entry valid.