core: set NoNewPrivileges in system mode when using seccomp
We need to also ensure that NoNewPrivileges is set even in system mode
when SystemCall* directives are used, and not only for user mode.

First, for security reasons it is important to set NoNewPrivileges when
using seccomp so that a malicious program could not call setuid programs
to gain new privileges.

Second, the manpage of seccomp specify that using seccomp with
SECCOMP_SET_MODE_FILTER will return EACCES if the caller does not have
CAP_SYS_ADMIN set, or if the no_new_privileges bit is not set. Hence,
without NoNewPrivileges set, it is impossible to use a SystemCall*
directive with a User directive set in system mode, or CAP_SYS_ADMIN
removed.