journal: avoid endless loop with --list-boot
The expectation that we will have entered a new boot after taking
one step does not hold in practice. After the seek with a match on
the previous boot id we should be at the boundary to the next boot
but after a next/previous step we sometimes remain within the same
boot. This leads to an endless loop over the same boot id. The
problem is only seen in 1 of the 200+ boots in my journal so it
seems likely that it is caused by corruption of that specific boot.

This patch matches the boot ids for the current and next journal
entry and keeps stepping forward until we are sure we are in a
boot.