siphash24: handle unaligned inputs
The siphash24 implementation assumes the input can be cast to a 64-bit integer.
This is not guaranteed to be true, when the input can be anything - including a
pointer into a string, which is the case for some of systemd's uses.

Adjust test-siphash24.c to call siphash24_compress() for different alignments,
to reproduce this reliably in the affected test.

Fix siphash24_compress() to copy the buffer onto a properly aligned pointer if
it is unaligned.