mount: always stop mount units after the local-fs.target unit during shutdown
For mount units generated by the fstab-generator, the ordering dep
"Before=local-fs.target' was conditionnaly added depending on some
options such as 'nofail' or 'noauto'.

I don't see any reason to not add it when 'noauto' is passed.

For the 'nofail' case, this allowed the local-fs.target unit to become
active without the need to wait for the mount unit to timeout when the
associated device node was missing.

Removing this ordering dep had an unwanted side effect however: during
shutdown, the affected mount units were not ordered against
local-fs.target anymore and could be stopped at any times, specially
before service units.

This patch unconditionnaly set "Before=local-fs.target" and order all
mount units to be stopped *after* local-fs.target is.

The side effect is that the boot process will wait for the timeout to
expire for missing disks even if the corresponding mount units have
'nofail' set.

A following patch will restore the correct behaviour regarding
'nofail' by setting the smallest possible timeout value for the device
units.